Fix
Fix for missing file
Fix for incorrect folder
Fix build error


Former-commit-id: 269e6a2d0e64f020cde31fe121f5098d3b929371 [formerly 2496eaec93a470be9d7852a07f0f64e38a22affa]
Former-commit-id: 727d1e46e125e16a44548a0bf994171f9cb682bb [formerly c24b4b82e30b6eb0c8f5b4bb35b1cf23cc650122]
Former-commit-id: 18823da25776bfe78f8b93cf80d72284251ba4f8
Former-commit-id: 03519d1bab26a1f1039ea3385e13ebdb6a6609c3

+++ Source/RawParser/Model/Effect/Demosaic.cs
@@ -24,7 +24,7 @@ namespace RawEditor.Effect
                     break;
                 case demosAlgorithm.NearNeighbour:
                 default:
-                    Demosaic.NearNeighbour(ref image, (int)image.dim.y, (int)image.dim.x, image.colorDepth, image.cfa);
+                    Demosaic.NearNeighbour(ref image, image.dim.y, image.dim.x, image.colorDepth, image.cfa);
                     break;
             }
         }
